// front-end/src/type/products/product-card-mini.interface.ts
/**
 * @description Product cart mini
 * @author Nhut Tan
 * @since 2025-09-28
 * @version 1.0.0
 */

/**
 * @property {string} name - Name of product
 * @property {string} imageUrl - Image url of product
 * @property {string} price - Price of product
 */
export interface ProductCardMiniInterface {
	name: string;
	imageUrl: string;
	price: string;
}